Cypress/E2E: Separated bot accounts in lists spec (#7470)
