+++ e2e/cypress/integration/bot_accounts/in_lists_1_spec.js
@@ -9,25 +9,14 @@
 
 // Group: @bot_accounts
 
-import {zip, sortBy} from 'lodash';
-
 import {createBotPatch} from '../../support/api/bots';
 import {generateRandomUser} from '../../support/api/user';
 
 describe('Bots in lists', () => {
     let team;
     let channel;
-    let testUser;
     let bots;
     let createdUsers;
-
-    const STATUS_PRIORITY = {
-        online: 0,
-        away: 1,
-        dnd: 2,
-        offline: 3,
-        ooo: 3,
-    };
 
     before(() => {
         cy.apiUpdateConfig({
@@ -38,7 +27,6 @@ describe('Bots in lists', () => {
         cy.apiInitSetup().then((out) => {
             team = out.team;
             channel = out.channel;
-            testUser = out.user;
         });
 
         cy.makeClient().then(async (client) => {
@@ -85,48 +73,6 @@ describe('Bots in lists', () => {
 
                 // * Verify pseudo checksum total of non bot users
                 cy.get('#searchableUserListTotal').should('have.text', `0 users of ${nonBotCount} total`);
-            });
-        });
-    });
-
-    it('MM-T1835 Channel Members list for BOTs', () => {
-        cy.makeClient({user: testUser}).then((client) => {
-            // # Login as regular user and visit a channel
-            cy.apiLogin(testUser);
-            cy.visit(`/${team.name}/channels/${channel.name}`);
-
-            // # Open channel members
-            cy.get('button.member-popover__trigger').click();
-
-            cy.get('#member-list-popover .more-modal__row .more-modal__name').then(async ($query) => {
-                // # Extract usernames from jQuery collection
-                const usernames = $query.toArray().map(({innerText}) => innerText);
-
-                // # Get users
-                const profiles = await client.getProfilesByUsernames(usernames);
-                const statuses = await client.getStatusesByIds(profiles.map((user) => user.id));
-                const users = zip(profiles, statuses).map(([profile, status]) => ({...profile, ...status}));
-
-                // # Sort 'em
-                const sortedUsers = sortBy(users, [
-                    ({is_bot: isBot}) => (isBot ? 1 : 0), // users first
-                    ({status}) => STATUS_PRIORITY[status],
-                    ({username}) => username,
-                ]);
-
-                // * Verify order of member-dropdown users against API-sourced/data-sorted version
-                cy.wrap(usernames).should('deep.equal', sortedUsers.map(({username}) => username));
-            });
-
-            // * Verify no statuses on bots
-            cy.get('#member-list-popover .more-modal__row--bot .status-wrapper .status').should('not.exist');
-
-            // * Verify bot badges
-            cy.get('#member-list-popover .more-modal__row--bot .Badge').then(($badges) => {
-                $badges.toArray().forEach((badgeEl) => {
-                    cy.wrap(badgeEl).then(() => badgeEl.scrollIntoView());
-                    cy.wrap(badgeEl).should('be.visible').and('have.text', 'BOT');
-                });
             });
         });
     });
